# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "httpx.ErrorDetail": {
                "properties": {
                    "code": {
                        "example": "invalid_request",
                        "type": "string"
                    },
                    "message": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "message"
                ],
                "type": "object"
            },
            "httpx.ErrorResponse": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/httpx.ErrorDetail"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
            "monitor.createRequest": {
                "properties": {
                    "interval_seconds": {
                        "example": 60,
                        "maximum": 2147483647,
                        "minimum": 1,
                        "type": "integer"
                    },
                    "url": {
                        "example": "https://example.com",
                        "type": "string"
                    }
                },
                "required": [
                    "interval_seconds",
                    "url"
                ],
                "type": "object"
            },
            "monitor.listResponse": {
                "properties": {
                    "monitors": {
                        "items": {
                            "$ref": "#/components/schemas/store.Monitor"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "next_cursor": {
                        "type": "string"
                    }
                },
                "required": [
                    "monitors"
                ],
                "type": "object"
            },
...
            "store.Monitor": {
                "properties": {
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "interval_seconds": {
                        "maximum": 2147483647,
                        "minimum": 1,
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "required": [
                    "created_at",
                    "id",
                    "interval_seconds",
                    "url"
                ],
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "Access JWT issued by registration, login, or refresh.",
                "scheme": "bearer",
                "type": "http"
            },
            "RefreshCookie": {
                "description": "HttpOnly refresh cookie scoped to /api/v1/auth. Sessions expire 30 days after login; rotation does not extend expiry.",
                "in": "cookie",
                "name": "refresh_token",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Authentication, profiles, avatars, and website monitor configuration. Unknown JSON fields are rejected. Monitoring checks are not implemented.\nRequests with Origin must match ALLOWED_ORIGIN. Mutating requests without Origin require X-CSRF-Protection: 1. Access JWTs expire after 900 seconds; logout does not invalidate existing JWTs.",
        "title": "Uptime API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/monitors": {
            "get": {
                "description": "Returns at most 50 monitors owned by the Bearer JWT identity, newest first by (created_at, id). Use next_cursor for the next page; it is omitted on the last page. Empty monitors is an array, not null.",
                "operationId": "listMonitors",
                "parameters": [
                    {
                        "description": "Opaque next_cursor from the preceding page",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "maxLength": 128,
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/monitor.listResponse"
                                }
                            }
                        },
                        "description": "Monitor page"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid page cursor"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unable to load monitors"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List website monitors",
                "tags": [
                    "monitors"
                ]
            },
            "post": {
                "description": "Saves configuration only; no outbound checks run. Owner is taken from the Bearer JWT. URL is trimmed and must be absolute HTTP/HTTPS without credentials, whitespace, or fragments, with an ASCII host and at most 2048 bytes. Interval must be a whole number of seconds.",
                "operationId": "createMonitor",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/monitor.createRequest",
                                "summary": "monitor",
                                "description": "Website URL and interval in seconds"
                            }
                        }
                    },
                    "description": "Website URL and interval in seconds",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/store.Monitor"
                                }
                            }
                        },
                        "description": "Created monitor"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid JSON, URL, or interval"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unable to create monitor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a website monitor",
                "tags": [
                    "monitors"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Current API server",
            "url": "/api/v1"
        }
    ]
}